+++ src/Gen3/RNGInfo.js
