+++ wasm/src/lib.rs
@@ -7,8 +7,8 @@ extern crate console_error_panic_hook;
 
 mod bdsp;
 mod enums;
+mod gen3;
 mod rng;
-mod rs;
 mod swsh;
 
 #[wasm_bindgen]
@@ -28,9 +28,9 @@ fn calculate_shiny_value(first: u16, second: u16) -> u16 {
 #[wasm_bindgen]
 pub fn get_gen3_wild(settings: &JsValue) -> JsValue {
     init_panic_hook();
-    let parsed_settings: rs::settings::Settings = settings.into_serde().unwrap();
+    let parsed_settings: gen3::settings::Settings = settings.into_serde().unwrap();
 
-    let results = rs::generator::generate_wild(parsed_settings);
+    let results = gen3::generator::generate_wild(parsed_settings);
 
     JsValue::from_serde(&results).unwrap()
 }
